modify: user auth and add db collection get to each auth file
